feat(core): added json helpers, directory helpers changes around making jsons and expanding directories.
