+++ uiautomationtools/helpers/dictionary_helpers.py
@@ -1,28 +1,5 @@
-import json
 import asyncio
 import pypeln as pl
-
-
-def deserialize(text, other_exceptions=None):
-    """
-    This converts json to its pythonic object.
-
-    Args:
-        text (str): The string to convert.
-        other_exceptions: Other exceptions to catch.
-
-    Returns:
-        data: The converted text or original text.
-    """
-    other_exceptions = other_exceptions or []
-    if type(other_exceptions) is not list:
-        other_exceptions = [other_exceptions]
-    exceptions = [json.JSONDecodeError] + other_exceptions
-
-    try:
-        return json.loads(text)
-    except tuple(exceptions):
-        return text
 
 
 def flatten(d):
